auth update
