--- apigateway/src/main/java/com/apigateway/WebConfig.java
@@ -1,18 +1,0 @@
-/*
- * package com.microservices.apigateway;
- * 
- * import org.springframework.context.annotation.Configuration; import
- * org.springframework.web.servlet.config.annotation.CorsRegistry; import
- * org.springframework.web.servlet.config.annotation.EnableWebMvc; import
- * org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
- * 
- * @Configuration
- * 
- * @EnableWebMvc public class WebConfig implements WebMvcConfigurer {
- * 
- * @Override public void addCorsMappings(CorsRegistry registry) {
- * registry.addMapping("*") .allowedOrigins("*") .allowedMethods("PUT",
- * "DELETE","POST","GET") .allowedHeaders("header1", "header2", "header3")
- * .exposedHeaders("header1", "header2") .allowCredentials(false).maxAge(3600);
- * } }
- */
