--- authentication/src/main/java/com/authentication/service/ProductService.java
@@ -1,50 +1,0 @@
-package com.authentication.service;
-
-import java.util.List;
-import java.util.Random;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-
-import com.authentication.dto.Product;
-import com.authentication.entity.UserInfo;
-import com.authentication.repository.UserInfoRepository;
-
-import jakarta.annotation.PostConstruct;
-
-@Service
-public class ProductService {
-
-	List<Product> productList = null;
-
-	@Autowired
-	private UserInfoRepository repository;
-
-	@Autowired
-	private PasswordEncoder passwordEncoder;
-
-	@PostConstruct
-	public void loadProductsFromDB() {
-		productList = IntStream.rangeClosed(1, 100)
-				.mapToObj(i -> new Product(i, "Product " + i, new Random().nextInt(10), new Random().nextInt(5000)))
-				.collect(Collectors.toList());
-	}
-
-	public List<Product> getProducts() {
-		return productList;
-	}
-
-	public Product getProduct(int id) {
-		return productList.stream().filter(product -> product.getProductId() == id).findAny()
-				.orElseThrow(() -> new RuntimeException("product " + id + " not found"));
-	}
-
-	public String addUser(UserInfo userInfo) {
-		userInfo.setPassword(passwordEncoder.encode(userInfo.getPassword()));
-		repository.save(userInfo);
-		return "BLogic added to system ";
-	}
-}
