+++ authentication/src/main/java/com/authentication/service/UserAuthService.java
@@ -1,0 +1,26 @@
+package com.authentication.service;
+
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+
+import com.authentication.entity.UserInfo;
+import com.authentication.repository.UserInfoRepository;
+
+
+@Service
+public class UserAuthService {
+
+	@Autowired
+	private UserInfoRepository repository;
+
+	@Autowired
+	private PasswordEncoder passwordEncoder;
+
+	public String addUser(UserInfo userInfo) {
+		userInfo.setPassword(passwordEncoder.encode(userInfo.getPassword()));
+		repository.save(userInfo);
+		return "BLogic added to system ";
+	}
+}
